rename: 컴포넌트 폴더 구조 세분화
기존 컴포넌트의 구조를 개선해 공통된 특성을 지닌 것들끼리 분류
Related to: #46
