+++ src/app/(dashboard)/dashboard/write/page.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState, Suspense } from 'react';
 import { useRouter, useSearchParams } from 'next/navigation';
 import Image from 'next/image';
 import Tiptap from '@/src/app/components/Tiptap';
-import Modal from '@/src/app/components/Modal';
+import Modal from '@/src/app/components/ui/Modal';
 import {
 	updatePostData,
 	createPostData,
