+++ src/app/(login)/layout.tsx
@@ -1,7 +1,7 @@
 'use client';
 import localFont from 'next/font/local';
 import '@/src/app/globals.css';
-import Providers from '../components/Providers';
+import Providers from '../components/layout/Providers';
 
 const pretendard_var = localFont({
 	src: '../../../public/fonts/pretendard_variable.woff2',
