+++ src/app/components/layout/SideNav.tsx
@@ -2,7 +2,7 @@ import NavLinks from './NavLinks';
 import Link from 'next/link';
 import { signout } from '@/src/utils/supabase/serverActions';
 import { PowerIcon } from 'lucide-react';
-import HoonieLogo from './icon/HoonieLogo';
+import HoonieLogo from '../icon/HoonieLogo';
 
 export default function SideNav() {
 	return (
